more trimming
