+++ funticula/funx/kelvin_effect.py
