+++ funticula/funx/mean_free_path.py
@@ -7,5 +7,5 @@ def molecule_mean_free_path(
     """ https://en.wikipedia.org/wiki/Mean_free_path """
     return (
         (2 * dynamic_viscosity / pressure)
-        / (8 * molar_mass / (pi * gas_constant * temperature))**0.5,
+        / (8 * molar_mass / (pi * gas_constant * temperature))**0.5
     )
