--- funticula/test/test_repo.py
@@ -1,8 +1,0 @@
-""" test """
-
-import funticula
-
-
-def test_funticula():
-    """ test funticula """
-    assert funticula.__version__ == '0.0.0'
